updating to current pull

+++ server/db/models/order.js
@@ -3,8 +3,8 @@ const db = require('../db');
 
 const Order = db.define('order', {
   totalCost: {
-    type: Sequelize.DOUBLE,
-    defaultValue: 0.0,
+    type: Sequelize.INTEGER,
+    defaultValue: 0,
   },
   status: {
     type: Sequelize.ENUM('open', 'closed'),
